Targeting VS2019 build. sentence.cpp updated.

+++ TextParser/TextParser/sentence.h
@@ -14,7 +14,7 @@ public:
 	int get_word_count();
 	bool is_interrogative_sentence();
 	std::vector<word*>* get_all_unique_words(int length);
-	void replace_all_words_by_length_with_substring(int word_length, std::vector<sentence_item*>* items);
+	void replace_all_words_by_length_with_substring(unsigned long word_length, std::vector<sentence_item*>* items);
 	void replace_all_sentence_words_start_with_consonant(int length);
 	std::string* get_sentence();
 	~sentence();
